# shared/swaggerCategory.yaml
openapi: 3.0.0
info:
  title: Category API
  version: 1.0.0
  description: API for managing categories.
servers:
  - url: http://localhost:3000/api
paths:
  /categories:
    get:
      summary: Retrieve all categories
      responses:
        '200':
          description: A list of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
        '500':
          description: Internal Server Error
    post:
      summary: Create a new category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  /categories/{id}:
    get:
      summary: Retrieve a category by ID
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the category to retrieve
          schema:
            type: string
      responses:
        '200':
          description: A single category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          description: Category not found
        '500':
          description: Internal Server Error
    put:
      summary: Update a category by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the category to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryInput'
      responses:
        '200':
          description: Updated category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad Request
        '404':
          description: Category not found
        '500':
          description: Internal Server Error
    delete:
      summary: Delete a category by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the category to delete
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Category not found
        '500':
          description: Internal Server Error

components:
  schemas:
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
      required:
        - id
        - name
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name